tone analyzer module functions with rest of app
